+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## XX.XX.XX
+* Added a new config option disableViewRestartForManualRecording to disable auto close/restart behavior of manual views on app background/foreground actions.
+
 ## 25.4.8
 * Mitigated an issue where "giveAllConsent" did not include metrics consent.